running unit tests for server logic in ci
