+++ webapp/vitest.server.config.ts
@@ -1,0 +1,12 @@
+import tsconfigPaths from "vite-tsconfig-paths";
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  plugins: [tsconfigPaths()],
+  test: {
+    globals: true,
+    environment: "node",
+    setupFiles: [],
+    include: ["app/test/server/*.{test,spec}.{ts,tsx}"],
+  },
+});
